demo/metaflow-demo/requirements.txt
metaflow
pyyaml
kubernetes